+++ src/components/index.ts
@@ -1,3 +1,4 @@
 export * from './component';
 export * from './metadata';
 export * from './schema';
+export * from './utils';