+++ src/index.ts
@@ -1,3 +1,4 @@
+export { isInstanceOf } from './components';
+export { component, computed, noDispatch, sequence, withId } from './decorators';
 export { SchemaOptions, GlobalOptions, LogLevel } from './options';
 export { ReduxApp } from './reduxApp';
-export { component, computed, noDispatch, sequence, withId } from './decorators';