+++ test/playground.ts
@@ -1,23 +1,21 @@
 import { expect } from 'chai';
-import { component } from 'src';
+import { component, isInstanceOf } from 'src';
 import { Component } from 'src/components';
 import { FakeStore } from './testTypes';
+
+// tslint:disable:no-unused-expression
 
 describe("playground", () => {
     it("it is a place to run wild", () => {
 
-        debugger;
-
         @component
-        class TheComponent {
-            public foo() {
-                return 5;
-            }
+        class MyComponent {
+            public value = 'hi';
         }
-        
-        var comp1: TheComponent = Component.create(new FakeStore(), new TheComponent()) as any;
-        var comp2: TheComponent = Component.create(new FakeStore(), new TheComponent()) as any;
 
-        expect(comp1.foo).to.equal(comp2.foo);
+        const myComponent = Component.create(new FakeStore(), new MyComponent());
+
+        debugger;
+        expect(isInstanceOf(myComponent, MyComponent)).to.be.true;
     });
 });